extract pm quality summary routes
