+++ docs/architecture/CODEBASE-REVIEW-LEDGER.md
@@ -2327,3 +2327,25 @@ This ledger records cleanup and structural review evidence for RFC-0036.
   score-run helper/core-resolver test seam explicit before moving score-run routes.
 - Wiki decision: no wiki source change required; this is internal route modularity cleanup with no
   route, payload, supported-feature, or operator-contract change.
+
+## BACKEND-REVIEW-20260531-076: PM operating quality summary-invocation routes were mixed with review flows
+
+- Date: 2026-05-31
+- Scope: PM operating quality support-summary invocation preview, create, list, detail, and
+  route-local builder behavior.
+- Finding: support-summary invocation endpoints and their validation builder lived in the parent
+  PM operating quality router alongside score-run, fairness, review-action, and policy routes.
+  These routes own append-only support-summary workflow evidence and have a distinct governance
+  boundary from supervisory review actions.
+- Action: moved the summary-invocation route group and builder into
+  `src/api/routers/pm_operating_quality_summary_routes.py` and mounted it from the parent PM
+  operating quality router, preserving public paths, response contracts, descriptions, conflict
+  handling, missing-target behavior, and validation error mapping.
+- Status: hardened
+- Evidence: PM operating quality API regression test (`tests/unit/api/test_pm_operating_quality_api.py`),
+  focused Ruff checks, source-file mypy, OpenAPI quality gate, and API vocabulary inventory
+  validation with no drift.
+- Follow-up: isolate review-action routes after making the parent-level builder monkeypatch
+  contract explicit, then finish score-run extraction.
+- Wiki decision: no wiki source change required; this is internal route modularity cleanup with no
+  route, payload, supported-feature, or operator-contract change.
